fix: format code
Main Changes:
1. move class NestedIterator from HugeTraverser.java to separate file
2. rollback to origin code
